+++ .flake8
@@ -27,6 +27,6 @@ max-line-length = 120
 max-complexity = 20
 import-order-style = pycharm
 application-import-names =
-    moltda
+    moleculetda
     tests
 format = ${cyan}%(path)s${reset}:${yellow_bold}%(row)d${reset}:${green_bold}%(col)d${reset}: ${red_bold}%(code)s${reset} %(text)s
