+++ .github/workflows/release.yml
@@ -12,4 +12,4 @@ jobs:
           token: ${{ secrets.GITHUB_TOKEN }}
 
           release-type: python
-          package-name: moltda
+          package-name: moleculetda
