+++ docs/Makefile
@@ -3,7 +3,7 @@
 # You can set these variables from the command line.
 SPHINXOPTS    =
 SPHINXBUILD   = sphinx-build
-SPHINXPROJ    = moltda
+SPHINXPROJ    = moleculetda
 SOURCEDIR     = source
 BUILDDIR      = build
