+++ docs/source/installation.rst
@@ -1,22 +1,22 @@
 Installation
 ============
 The most recent release can be installed from
-`PyPI <https://pypi.org/project/moltda>`_ with:
+`PyPI <https://pypi.org/project/moleculetda>`_ with:
 
 .. code-block:: shell
 
-    $ pip install moltda
+    $ pip install moleculetda
 
 The most recent code and data can be installed directly from GitHub with:
 
 .. code-block:: shell
 
-    $ pip install git+https://github.com/kjappelbaum/moltda.git
+    $ pip install git+https://github.com/kjappelbaum/moleculetda.git
 
 To install in development mode, use the following:
 
 .. code-block:: shell
 
-    $ git clone git+https://github.com/kjappelbaum/moltda.git
-    $ cd moltda
+    $ git clone git+https://github.com/kjappelbaum/moleculetda.git
+    $ cd moleculetda
     $ pip install -e .
