+++ setup.cfg
@@ -1,6 +1,6 @@
 
 [metadata]
-name = moltda
+name = moleculetda
 version = 0.1.0
 description = Toolkit for using topological data analysis representations.
 author = Aditi Krishnapriyan 
@@ -67,7 +67,7 @@ where = src
 
 [options.entry_points]
 console_scripts =
-    moltda = moltda.cli:main
+    moleculetda = moleculetda.cli:main
 
 ######################
 # Doc8 Configuration #
@@ -82,15 +82,15 @@ max-line-length = 120
 ##########################
 [coverage:run]
 branch = True
-source = moltda
+source = moleculetda
 omit =
     tests/*
     docs/*
 
 [coverage:paths]
 source =
-    src/moltda
-    .tox/*/lib/python*/site-packages/moltda
+    src/moleculetda
+    .tox/*/lib/python*/site-packages/moleculetda
 
 [coverage:report]
 show_missing = True
