+++ src/moleculetda/io.py
@@ -1,7 +1,7 @@
 import json
+import pickle
 
 import numpy as np
-import pickle
 
 
 class NumpyEncoder(json.JSONEncoder):
