+++ src/moleculetda/read_file.py
@@ -5,8 +5,8 @@ from pathlib import Path
 from typing import Tuple, Union
 
 import numpy as np
-from pymatgen.core import Molecule, Structure
 from loguru import logger
+from pymatgen.core import Molecule, Structure
 from pymatgen.transformations.advanced_transformations import CubicSupercellTransformation
