+++ src/moleculetda/structure_to_vectorization.py
