+++ tox.ini
@@ -42,8 +42,8 @@ deps =
     isort
 skip_install = true
 commands =
-    black src/moltda/ tests/ setup.py
-    isort src/moltda/ tests/ setup.py
+    black src/moleculetda/ tests/ setup.py
+    isort src/moleculetda/ tests/ setup.py
 description = Run linters.
 
 [testenv:manifest]
@@ -67,7 +67,7 @@ deps =
     pep8-naming
     pydocstyle
 commands =
-    flake8 src/moltda/ tests/ setup.py
+    flake8 src/moleculetda/ tests/ setup.py
 description = Run the flake8 tool with several plugins (bandit, docstrings, import order, pep8 naming). See https://cthoyt.com/2020/04/25/how-to-code-with-me-flake8.html for more information.
 
 [testenv:pyroma]
@@ -81,7 +81,7 @@ description = Run the pyroma tool to check the package friendliness of the proje
 [testenv:mypy]
 deps = mypy
 skip_install = true
-commands = mypy --install-types --non-interactive --ignore-missing-imports src/moltda/
+commands = mypy --install-types --non-interactive --ignore-missing-imports src/moleculetda/
 description = Run the mypy tool to check static typing on the project.
 
 [testenv:doc8]
@@ -98,7 +98,7 @@ skip_install = true
 deps =
     docstr-coverage
 commands =
-    docstr-coverage src/moltda/ tests/ setup.py --skip-private --skip-magic
+    docstr-coverage src/moleculetda/ tests/ setup.py --skip-private --skip-magic
 description = Run the docstr-coverage tool to check documentation coverage
 
 [testenv:docs]
